[ExchangeWithWS] fix order resync
missed the recently executed ones (sync only got the active orders). so could lead to problem with sync removing the executed before the bot gets the info that it was exectued.
added the recently ones back to the list. also added reset of timer to not resync orders right after something changed there.